+++ src/components/game-artist-page/game-artist-page.jsx
