+++ src/components/game-genre-page/game-genre-page.jsx
