+++ src/components/welcome-page/welcome-page.jsx
